+++ src/main/java/com/curioussong/alsongdalsong/song/repository/SongRepository.java
@@ -10,5 +10,5 @@ import java.util.List;
 public interface SongRepository extends JpaRepository<Song, Long> {
 
     @Query(value = "SELECT * FROM song WHERE year IN (:years) ORDER BY RAND() LIMIT :maxRound", nativeQuery = true)
-    List<Song> findRandomSongsByYears(@Param("years") List<Long> years, @Param("maxRound") Integer maxRound);
+    List<Song> findRandomSongsByYears(@Param("years") List<Integer> years, @Param("maxRound") Integer maxRound);
 }
